Support 3 states of help: hidden, visible, full
After a couple usages the keys should be memorized, so the users might
want to hide the help completely to add another row to the code
view/list.

In the future, help section state will be persisted in the optional
configuration file, so that for every new gocovsh session the menu will
be in the selected state.