Support explicit selection of files to display
In larger projects, running gocovsh after making a small change will
cause redundant work to find the changed files in the long, multi-page
list. This flow appears to be the most common for a routine work:
usually the developer will want to view their change coverage, and not
the entire project coverage.

To achieve this goal, the list of files should be piped into gocovsh. It
allows to keep the "list" operation abstract: it can be a git diff with
any arguments the user desires, if the result is a simple multi-line
list of files.

For example, the suggested usage would be:
git diff --name-only | gocovsh		# view non-stashed changes
git diff --name-only master | gocovsh	# view changes relative to master